Fix handling of multiple build tags
Previously, each build tag was being passed to `go list` as a separate
argument, which caused `go list` to interpret the second one as a
package name, which it couldn't find. This meant that bingo only worked
with zero or one build tags, not more than one.

Join the build tags with spaces and pass them all as a single argument
to `go list`.